Docs: describe low- and high-level components, add low-level ccall example
Describe in the README.md and the front page of the documentation how
PETSc.jl includes a lower-level automated layer which wraps the PETSc C
API, along with a (smaller) set of (volatile) high-level functions.

For PETSc experts wanting to use additional functions or contribute new
wrappers, add a minimal example of using ccall to call a low-level
function directly.